requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20210324.1=h0e60522_0
alabaster=0.7.12=py37_0
altair=4.1.0=py_1
argcomplete=1.12.3=pyhd3eb1b0_0
argh=0.26.2=pyh9f0ad1d_1002
argon2-cffi=20.1.0=py37h2bbff1b_1
arrow-cpp=4.0.0=py37hb1a8454_3_cpu
astor=0.8.1=pyh9f0ad1d_0
astroid=2.6.6=py37haa95532_0
async_generator=1.10=py37h28b3542_0
atomicwrites=1.4.0=pypi_0
attrs=21.2.0=pyhd3eb1b0_0
aws-c-cal=0.5.11=he19cf47_0
aws-c-common=0.6.2=h8ffe710_0
aws-c-event-stream=0.2.7=h70e1b0c_13
aws-c-io=0.10.5=h2fe331c_0
aws-checksums=0.1.11=h1e232aa_7
aws-sdk-cpp=1.8.186=hb0612c5_3
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.0=py_2
backports.zoneinfo=0.2.1=py37hcc03f2d_4
base58=2.1.0=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blinker=1.4=py_1
boto3=1.18.54=pyhd8ed1ab_0
botocore=1.21.54=pyhd8ed1ab_0
bottleneck=1.3.2=py37h2a96729_1
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37hcc03f2d_1001
bzip2=1.0.8=h8ffe710_4
c-ares=1.17.2=h8ffe710_0
ca-certificates=2021.7.5=haa95532_1
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.14.12=hf171d8a_3
certifi=2021.5.30=py37haa95532_0
cffi=1.14.6=py37h2bbff1b_0
chardet=4.0.0=py37h03978a9_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
chembl_webresource_client=0.10.1=pyh9f0ad1d_0
click=7.1.2=pyh9f0ad1d_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.7=py37h20c650d_0
cycler=0.10.0=py37_0
debugpy=1.4.1=py37hd77b12b_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docutils=0.17.1=py37haa95532_1
easydict=1.9=py_0
entrypoints=0.3=py37_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
gevent=21.8.0=py37hcc03f2d_0
gflags=2.2.2=ha925a31_1004
gitdb=4.0.7=pyhd8ed1ab_0
gitpython=3.1.24=pyhd8ed1ab_0
glog=0.5.0=h4797de2_0
greenlet=1.1.1=py37hf2a7229_0
grequests=0.6.0=pyh9f0ad1d_0
grpc-cpp=1.37.1=h586195c_2
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.1=pyhd3deb0d_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=4.8.1=py37haa95532_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=haa95532_3372
ipykernel=6.2.0=py37haa95532_1
ipython=7.27.0=py37hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
isort=5.9.3=pyhd3eb1b0_0
jedi=0.18.0=py37haa95532_1
jinja2=3.0.1=pyhd3eb1b0_0
jmespath=0.10.0=pyh9f0ad1d_0
joblib=1.0.0=pypi_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py37_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keyring=23.1.0=py37haa95532_0
kiwisolver=1.3.1=py37hd77b12b_0
krb5=1.19.2=hbae68bd_0
lazy-object-proxy=1.6.0=py37h2bbff1b_0
lazypredict=0.2.9=pypi_0
libboost=1.67.0=hd9e427e_4
libcurl=7.79.1=h789b8ee_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.16.0=h7755175_0
libssh2=1.10.0=h680486a_2
libthrift=0.14.1=h636ae23_2
libtiff=4.2.0=hd0e1b90_0
libutf8proc=2.6.1=hcb41399_0
lightgbm=2.3.1=pypi_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=py37h2bbff1b_0
matplotlib=3.4.2=py37haa95532_0
matplotlib-base=3.4.2=py37h49ac443_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py37_1
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.0=py37h277e83a_2
mkl_random=1.2.2=py37hf11a4ad_0
more-itertools=8.10.0=pypi_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.3=py37haa95532_0
numexpr=2.7.3=py37hb80d3ca_1
numpy=1.19.1=pypi_0
numpy-base=1.20.3=py37hc2deb75_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=py37_0
openssl=1.1.1l=h2bbff1b_0
packaging=21.0=pyhd3eb1b0_0
padelpy=0.1.10=pypi_0
pandas=1.0.5=pypi_0
pandocfilters=1.4.3=py37haa95532_1
parquet-cpp=1.5.1=2
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py37h4fa10fc_0
pip=21.0.1=py37haa95532_0
pixman=0.40.0=h2bbff1b_1
pluggy=0.13.1=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyhca03da5_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.16.0=py37hf2a7229_0
psutil=5.8.0=py37h2bbff1b_1
py=1.10.0=pypi_0
py-boost=1.67.0=py37h8300f20_4
pyarrow=4.0.0=py37h0b73db8_3_cpu
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydeck=0.6.2=pyh44b312d_0
pyflakes=2.3.1=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pylint=2.9.6=py37haa95532_1
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37h03978a9_3
pytest=5.4.3=pypi_0
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2021.2.post0=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2021.1=pyhd3eb1b0_0
pywin32=228=py37hbaba5e8_1
pywin32-ctypes=0.2.0=py37_1001
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=pypi_0
pyzmq=22.2.1=py37hd77b12b_1
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.2=pyhd3eb1b0_0
qtconsole=5.1.0=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
rdkit=2020.09.1.0=py37h3d1ada6_1
re2=2021.04.01=h0e60522_0
requests=2.26.0=pyhd8ed1ab_0
requests_cache=0.4.13=py_0
rope=0.19.0=pyhd3eb1b0_0
s3transfer=0.5.0=pyhd8ed1ab_0
scikit-learn=0.23.1=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37haa95532_0
sip=4.19.8=py37h6538335_0
six=1.15.0=pypi_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.8=ha925a31_3
snowballstemmer=2.1.0=pyhd3eb1b0_0
sphinx=4.2.0=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=3.3.6=py37_0
spyder-kernels=0.5.2=py37_0
sqlite=3.36.0=h2bbff1b_0
streamlit=0.84.1=pyhd8ed1ab_0
terminado=0.9.4=py37haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.10=he774522_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.1=py_0
tornado=6.1=py37h2bbff1b_0
tqdm=4.56.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typed-ast=1.4.3=py37h2bbff1b_1
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021b=he74cb21_0
tzlocal=3.0=py37h03978a9_2
urllib3=1.26.7=pyhd8ed1ab_0
validators=0.18.2=pyhd3deb0d_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py37h03978a9_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
wget=1.16.3=0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py37_0
win_inet_pton=1.1.0=py37h03978a9_2
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xgboost=1.1.1=pypi_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zope.event=4.5.0=pyh9f0ad1d_0
zope.interface=5.4.0=py37hcc03f2d_0
zstd=1.4.9=h19a0ad4_0
